ADD simple log wrapper with DEBUG level option
the log doesn't misses the DEBUG level. I'm setting up
information in two groups INFO or DATA that will print
anyway with the difference is that info comes from the
app itself informing the user what's happening and DATA
is the data being retrieved by the app and serve on the
container or pod log output.

DEBUG can be activated by setting the environment var
PODTRACER_LOGLEVEL="DEBUG". Very low level detail will
be printed to the container or pod log.