ADD writers.go
The intention is to build all the writers used by
os/exec cmd using the same standard regardless on
what kind of target it will write to. If a file or
buffer or database. The data transport, connections
and implementation details for each target will
have their own methods setSomethingWritters to
return a writer that os/exec can write to through
io.MultiWriters.

Each method needs to append a closer function to
writers.Closers. That's used by the cleanup method
to gracefully shutdown all open connections.