REMOVE kubeconfig from flags
Kubeconfig was used in the beginning for tests. Podtracer
now uses the token assigned to the service account that runs it.

That said no logic was added to kubeconfig. Instead kubeconfig
was removed as an option for now. There may be a use case to
inject a kubeconfig flag which would be a remote usage, i. e.,
podtracer being executed in a remote laptop against a cluster
on the cloud but that a full feature story which is not the
case for now.